Move WS connection outside of Luigi + Refactoring
