+++ modules/luigi/__init__.py
@@ -4,8 +4,6 @@ import ravestate_interloc as interloc
 import ravestate_rawio as rawio
 import ravestate_idle as idle
 import os
-import rospy
-from roboy_cognition_msgs.srv import DriveToLocation, DriveToLocationResponse
 import actionlib
 #from roboy_cognition_msgs.msg import OrderIceCreamAction, OrderIceCreamGoal
 #from roboy_cognition_msgs.srv import Payment
@@ -13,14 +11,9 @@ from enum import IntEnum
 from ravestate_verbaliser import verbaliser
 from os.path import realpath, dirname, join
 verbaliser.add_folder(join(dirname(realpath(__file__))+"/phrases"))
-import asyncio
-import websockets
-import pickle
 
 cost_per_scoop = 1  # TODO move to external config file that also lists the available flavors and payment options
 
-location =""
-server = 'ws://localhost:8765'#TODO change to cloud
 
 FLAVORS = {"chocolate", "vanilla"}
 FLAVOR_SYNONYMS = {"flavor", "kind"}
@@ -258,23 +251,23 @@ with rs.Module(name="Luigi"):
 
     # -------------------- states: conversation flow -------------------- #
 
-    @rs.state(
-        cond=interloc.prop_all.pushed().detached().min_age(2) | idle.sig_bored.min_age(1),
-        signal=sig_send_eta,
-        emit_detached=True)
-    def get_loc_send_eta(ctx: rs.ContextWrapper):
-        communication_with_cloud(server)
-        return rs.Emit()
-
-
-    @rs.state(
-        cond=sig_send_eta,
-        signal=sig_has_arrived,
-        emit_detached=True)
-    def arrived_at_location(ctx: rs.ContextWrapper):
-        has_arrived = True #TODO change this with has_arrived from ros
-        if has_arrived:
-            return rs.Emit()
+    # @rs.state(
+    #     cond=interloc.prop_all.pushed().detached().min_age(2) | idle.sig_bored.min_age(1),
+    #     signal=sig_send_eta,
+    #     emit_detached=True)
+    # def get_loc_send_eta(ctx: rs.ContextWrapper):
+    #     communication_with_cloud(server)
+    #     return rs.Emit()
+    #
+    #
+    # @rs.state(
+    #     cond=sig_send_eta,
+    #     signal=sig_has_arrived,
+    #     emit_detached=True)
+    # def arrived_at_location(ctx: rs.ContextWrapper):
+    #     has_arrived = True #TODO change this with has_arrived from ros
+    #     if has_arrived:
+    #         return rs.Emit()
 
 
     @rs.state(
@@ -592,37 +585,3 @@ def payment_communication(price, payment_option):
     # time.sleep(4)
     # return 220, ""
 
-
-def ad_communication(location):
-    rospy.wait_for_service('autonomous_driving')
-    try:
-        drive_to_location = rospy.ServiceProxy('autonomous_driving', DriveToLocation)
-        response = drive_to_location(location)
-        return response.eta, response.error_message
-    except rospy.ROSInterruptException as e:
-        print('Service call failed:', e)
-    # If driving module is run without ROS, comment everything from above (including imports) and uncomment this:
-    return 41, ""
-
-def communication_with_cloud(server):
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-    asyncio.get_event_loop().run_until_complete(listen(server))
-    #print("Listening is finished")
-    eta, err = ad_communication(location)
-    #print("Ros communication is finished")
-    asyncio.get_event_loop().run_until_complete(say(server,eta))
-    #print("Sending eta is finished")
-    loop.close()
-
-
-async def say(server, eta):
-    async with websockets.connect(server+'/pub') as websocket:
-        eta_encoding = pickle.dumps(eta)
-        await websocket.send(eta_encoding)
-
-async def listen(server):
-    global location
-    async with websockets.connect(server+'/sub') as websocket:
-        location_encoding = await websocket.recv()
-        location = pickle.loads(location_encoding, encoding='bytes')  # .decode()
